third_party/blink/web_tests/wpt_internal/framework-detection/not-detected-bad-values.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

</script>
<body>
  <div id="__next"></div>
  <script>

// Version should be a string
window.Vue = {version: 3 };
window.__NEXT_DATA__ = {};

// next without "version" key
window.next = {ver_sion: '13.3.2-canary.11'};

promise_test(async t => {
  const recorder = internals.initializeUKMRecorder();
  await new Promise(resolve => window.addEventListener("load", resolve));
  await new Promise(resolve => t.step_timeout(resolve));
  const entries = recorder.getMetrics(
    "JavascriptFrameworkVersions",
    ["NextJSVersion", "AngularVersion", "NuxtVersion", "VueVersion"]);
  assert_equals(entries.length, 0);
}, "With semi-invalid indicators, no framework should be detected");
  </script>
</body>
